requirements.txt
numpy
pandas
matplotlib
numpy-stl
opencv-python
scikit-image
tripy
open3d
tensorflow==2.10.1
scikit-learn
streamlit
scikeras
optuna
numba
